moved bits/sample info into fileinfo in SignalManager, changed sample rate data type to double
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@648 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ kwave/PluginManager.cpp
@@ -454,7 +454,7 @@ unsigned int PluginManager::signalLength()
 }
 
 //***************************************************************************
-unsigned int PluginManager::signalRate()
+double PluginManager::signalRate()
 {
     return m_top_widget.signalManager().rate();
 }
